minor bug fixes and added continuity logic

+++ logic/zerodha_integration_management/dtos/timeframe_budget_details_dto.py
@@ -1,13 +1,12 @@
 from pydantic import BaseModel
 from typing import Optional
 
-from data.enums.budget_part import BudgetPart
 from data.enums.trades import Trades
 
 
 class TimeframeBudgetDetailsDTO(BaseModel):
     budget: int
     no_of_trades: Trades
-    start_range: Optional[int]
-    end_range: Optional[int]
+    start_range: Optional[float]
+    end_range: Optional[float]
